+++ ui/dashboards/src/context/DashboardProvider/index.ts
@@ -13,5 +13,5 @@
 
 export * from './dashboard-provider-api';
 export * from './DashboardProvider';
-export type { PanelGroupId, PanelGroupDefinition, PanelGroupItemId as LayoutItem } from './layout-slice';
-export type { PanelGroupEditor, PanelGroupEditorValues } from './panel-group-slice';
+export type { PanelGroupId, PanelGroupDefinition, PanelGroupItemId as LayoutItem } from './panel-group-slice';
+export type { PanelGroupEditor, PanelGroupEditorValues } from './panel-group-editor-slice';
