+++ ui/dashboards/src/context/DashboardProvider/panel-editor-slice.tsx
@@ -15,10 +15,15 @@ import { PanelDefinition, UnknownSpec } from '@perses-dev/core';
 import { StateCreator } from 'zustand';
 import { removeWhiteSpacesAndSpecialCharacters } from '../../utils/functions';
 import { Middleware } from './common';
-import { LayoutSlice, PanelGroupItemId, PanelGroupId } from './layout-slice';
+import { PanelGroupSlice, PanelGroupItemId, PanelGroupId } from './panel-group-slice';
 
+/**
+ * Slice that handles the visual editor state and actions for Panels (i.e. add, edit, delete).
+ */
 export interface PanelEditorSlice {
+  // TODO: Move panels state to its own slice
   panels: Record<string, PanelDefinition>;
+
   deletePanelDialog?: DeletePanelDialog;
   /**
    * State for the panel editor when its open, otherwise undefined when it's closed.
@@ -26,14 +31,14 @@ export interface PanelEditorSlice {
   panelEditor?: PanelEditorState;
 
   /**
-   * Edit an existing panel by providing its layout coordinates.
+   * Opens the editor for editing an existing panel by providing its layout coordinates.
    */
-  editPanel: (item: PanelGroupItemId) => void;
+  openEditPanel: (item: PanelGroupItemId) => void;
 
   /**
-   * Add a new Panel to a panel group.
+   * Opens the editor for adding a new Panel to a panel group.
    */
-  addPanel: (panelGroupId?: PanelGroupId) => void;
+  openAddPanel: (panelGroupId?: PanelGroupId) => void;
 
   /**
    * Delete panels
@@ -95,14 +100,14 @@ export interface PanelEditorValues {
  */
 export function createPanelEditorSlice(
   panels: PanelEditorSlice['panels']
-): StateCreator<PanelEditorSlice & LayoutSlice, Middleware, [], PanelEditorSlice> {
+): StateCreator<PanelEditorSlice & PanelGroupSlice, Middleware, [], PanelEditorSlice> {
   // Return the state creator function for Zustand that uses the panels provided as intitial state
   return (set, get) => ({
     panels,
 
     panelEditor: undefined,
 
-    editPanel(item) {
+    openEditPanel(item) {
       const { panels, getPanelKey } = get();
 
       // Ask the layout store for the panel key at that location
@@ -147,7 +152,7 @@ export function createPanelEditorSlice(
       });
     },
 
-    addPanel(panelGroupId) {
+    openAddPanel(panelGroupId) {
       // If a panel group isn't supplied, add to the first group
       if (panelGroupId === undefined) {
         const firstGroupId = get().panelGroupIdOrder[0];
